+++ packages/react/src/Krona.tsx
@@ -6,6 +6,7 @@ import { ExpandBar } from './parts/ExpandBar'
 import { Gutter } from './parts/Gutter'
 import { Lines } from './parts/Lines'
 import { Minimap } from './parts/Minimap'
+import { SideSwitch } from './parts/SideSwitch'
 import { Toolbar } from './parts/Toolbar'
 import { KronaViewer } from './viewer/Viewer'
 
@@ -55,4 +56,5 @@ export const Krona = Object.assign(KronaRoot, {
   Diagnostics,
   Minimap,
   ExpandBar,
+  SideSwitch,
 })